+++ lib/cgo/params.distribution.go
@@ -15,12 +15,6 @@ import (
 */
 import "C"
 
-//export SKY_params_GetDistributionAddresses
-func SKY_params_GetDistributionAddresses(_arg0 *C.GoSlice_) {
-	__arg0 := params.GetDistributionAddresses()
-	copyToGoSlice(reflect.ValueOf(__arg0), _arg0)
-}
-
 //export SKY_params_GetUnlockedDistributionAddresses
 func SKY_params_GetUnlockedDistributionAddresses(_arg0 *C.GoSlice_) {
 	__arg0 := params.GetUnlockedDistributionAddresses()