+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Removed function `SKY_api_Client_CreateUnencryptedWallet`
 - Removed function `SKY_api_Client_CreateEncryptedWallet`
 - Removed function `SKY_cli_GenerateWallet`
+- Removed function `SKY_params_GetDistributionAddresses`
 
 ### Changed